+++ bin/check-autoscaling-cpucredits.rb
@@ -85,8 +85,7 @@ class CheckEc2CpuCredits < Sensu::Plugin::Check::CLI
   def aws_config
     { access_key_id: config[:aws_access_key],
       secret_access_key: config[:aws_secret_access_key],
-      region: config[:aws_region]
-    }
+      region: config[:aws_region] }
   end
 
   def asg