+++ bin/check-s3-bucket.rb
@@ -62,8 +62,7 @@ class CheckS3Bucket < Sensu::Plugin::Check::CLI
   def aws_config
     { access_key_id: config[:aws_access_key],
       secret_access_key: config[:aws_secret_access_key],
-      region: config[:aws_region]
-    }
+      region: config[:aws_region] }
   end
 
   def run