+++ bin/metrics-ec2-count.rb
@@ -21,7 +21,7 @@
 # NOTES:
 #
 # LICENSE:
-#   Copyright (c) 2014, Tim Smith, tim@cozy.co
+#   Copyright (c) 2014, Tim Smith, tsmith@chef.io
 #   Released under the same terms as Sensu (the MIT license); see LICENSE
 #   for details.
 #
@@ -63,8 +63,7 @@ class EC2Metrics < Sensu::Plugin::Metric::CLI::Graphite
   def aws_config
     { access_key_id: config[:aws_access_key],
       secret_access_key: config[:aws_secret_access_key],
-      region: config[:aws_region]
-    }
+      region: config[:aws_region] }
   end
 
   def by_instances_status(client)