+++ bin/metrics-sqs.rb
@@ -70,8 +70,7 @@ class SQSMetrics < Sensu::Plugin::Metric::CLI::Graphite
   def aws_config
     { access_key_id: config[:aws_access_key],
       secret_access_key: config[:aws_secret_access_key],
-      region: config[:aws_region]
-    }
+      region: config[:aws_region] }
   end
 
   def scheme(queue_name)